% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoPoint.R
\docType{data}
\name{dot_data}
\alias{dot_data}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{data.frame} with 168 rows and 6 columns.
}
\usage{
dot_data
}
\description{
This is a test data for this package
test data describtion
}
\author{
Junjun Lao
}
\keyword{datasets}
