% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoRect.R
\docType{data}
\name{pgo}
\alias{pgo}
\title{This is a test data for this package
test data describtion}
\format{
An object of class \code{gg} (inherits from \code{ggplot}) of length 9.
}
\usage{
pgo
}
\description{
This is a test data for this package
test data describtion
}
\author{
Junjun Lao
}
\keyword{datasets}
