% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useMyCol.R
\name{useMyCol}
\alias{useMyCol}
\title{useMyCol}
\usage{
useMyCol(platte = NULL, n = NULL, showAll = FALSE)
}
\arguments{
\item{platte}{The platte name. Default("stallion").}

\item{n}{The color numbers to use. Default(NULL).}

\item{showAll}{Whether to show all plattes. Default(FALSE).}
}
\value{
Return the color names you have choosed.
}
\description{
This function is used to produce avaliable colors for plot.
}
\examples{
useMyCol(platte = 'stallion2',n = 5)
useMyCol(showAll = TRUE)
}
\author{
Junjun Lao
}
