% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoRect.R
\name{annoRect}
\alias{annoRect}
\title{annoRect}
\usage{
annoRect(
  object = NULL,
  annoPos = "top",
  xPosition = NULL,
  yPosition = NULL,
  pCol = NULL,
  pFill = NULL,
  rectWidth = 1,
  lty = NULL,
  lwd = NULL,
  alpha = NULL,
  annoManual = FALSE
)
}
\arguments{
\item{object}{Your ggplot list. Default(NULL).}

\item{annoPos}{The position for the annotation to be added. Default("top").}

\item{xPosition}{The x axis coordinate for the rect. Default(NULL).}

\item{yPosition}{The y axis coordinate for the rect. Default(NULL).}

\item{pCol}{The rect colors. Default(NULL).}

\item{pFill}{The rect fill colors. Default(NULL).}

\item{rectWidth}{The relative rect width. Default(1).}

\item{lty}{The rect line type. Default(NULL).}

\item{lwd}{The rect line width. Default(NULL).}

\item{alpha}{The rect fill color alpha. Default(NULL).}

\item{annoManual}{Whether annotate by yourself by supplying with x and y coordinates. Default(FALSE).}
}
\value{
Return a ggplot object.
}
\description{
This function is used to add rect annotations in plot.
}
\examples{
# ===============================
# test function

# load data
data(p)
data(pgo)

# default plot
annoRect(object = p,
         annoPos = 'top',
         xPosition = c(1:10))

# you can set y axis no expand
annoRect(object = p,
         annoPos = 'top',
         xPosition = c(1:10)) +
  ggplot2::scale_y_discrete(expand = c(0,0))

# adjust yPosition
annoRect(object = p,
         annoPos = 'top',
         xPosition = c(1:10),
         yPosition = c(11,11.5))

# another example annotation GO terms
annoRect(object = pgo,
        annoPos = 'right',
        yPosition = c(1:15),
        pCol = rep('transparent',15),
        pFill = rep(c('#F5F0BB','#C4DFAA','#90C8AC'),each = 5),
        xPosition = c(3,9.5),
        rectWidth = 1)
}
\author{
Junjun Lao
}
